+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmDAO.java
@@ -23,7 +23,7 @@ public interface FilmDAO {
 
     boolean deleteLike(Long filmId, Long userId);
 
-    List<Film> getPopularFilm(Integer count, Integer genreId, Integer year);
+    List<Film> getPopularFilm(Integer count, Long genreId, Integer year);
 
     List<Film> findAllFilmsByDirectorId(Long directorId, String sortBy);
 }
